BELA/requirements_bela.txt
faiss-gpu
fairscale
hydra-core
hydra-submitit-launcher
pyyaml
torch==2.0.1
pytorch-lightning
transformers
tqdm
sentencepiece
h5py
protobuf==3.20
ujson
